Added switches to turn on VPN and internet tunnel

+++ app/src/main/java/com/anode/anode/MainActivity.kt
@@ -3,60 +3,71 @@ package com.anode.anode
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
-import android.net.VpnService
 import android.os.Bundle
+import android.text.Html
+import android.text.Spanned
+import android.text.method.LinkMovementMethod
 import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
+import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
+import androidx.core.text.HtmlCompat
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.snackbar.Snackbar
+import org.w3c.dom.Text
 import java.io.*
+
 
 class MainActivity : AppCompatActivity() {
 
     private var util: AnodeUtil? = null
 
-    private fun InitializeApp() {
+    private fun initializeApp() {
         //Create files folder
         application.filesDir.mkdir()
-        //Copy cjdroute
-        var `in`: InputStream?
-        var out: OutputStream?
-        try {
-            val am = baseContext.assets
-            //Read architecture
-            val arch = System.getProperty("os.arch")
-            Log.i(LOGTAG, "OS Architecture: $arch")
-            `in` = if (arch!!.contains("x86") || arch.contains("i686")) {
-                am.open("x86/cjdroute")
-            } else if (arch.contains("arm64-v8a")) {
-                am.open("arm64-v8a/cjdroute")
-            } else if (arch.contains("armeabi")) {
-                am.open("armeabi-v7a/cjdroute")
-            } else { //Unknown architecture
-                Log.i(LOGTAG, "Incompatible CPU architecture")
-                return
+        val cjdrouteFile = File(AnodeUtil().CJDNS_PATH+"/"+AnodeUtil().cjdrouteConfFile)
+        if (!cjdrouteFile.exists()) {
+            //Copy cjdroute
+            var `in`: InputStream?
+            var out: OutputStream?
+            try {
+                val am = baseContext.assets
+                //Read architecture
+                val arch = System.getProperty("os.arch")
+                Log.i(LOGTAG, "OS Architecture: $arch")
+                `in` = if (arch!!.contains("x86") || arch.contains("i686")) {
+                    am.open("x86/cjdroute")
+                } else if (arch.contains("arm64-v8a") || arch.contains("aarch64")) {
+                    am.open("arm64-v8a/cjdroute")
+                } else if (arch.contains("armeabi")) {
+                    am.open("armeabi-v7a/cjdroute")
+                } else { //Unknown architecture
+                    Log.i(LOGTAG, "Incompatible CPU architecture")
+                    return
+                }
+                out = FileOutputStream(application.filesDir.toString() + "/cjdroute")
+                val buffer = ByteArray(1024)
+                var read: Int
+                while (`in`.read(buffer).also { read = it } != -1) {
+                    out.write(buffer, 0, read)
+                }
+                `in`.close()
+                out.close()
+                //Set permissions
+                val file = File(application.filesDir.toString() + "/cjdroute")
+                file.setExecutable(true)
+            } catch (e: IOException) {
+                e.printStackTrace()
+                Log.e(LOGTAG, "Failed to copy cjdroute file", e)
             }
-            out = FileOutputStream(application.filesDir.toString() + "/cjdroute")
-            val buffer = ByteArray(1024)
-            var read: Int
-            while (`in`.read(buffer).also { read = it } != -1) {
-                out.write(buffer, 0, read)
-            }
-            `in`.close()
-            out.close()
-            //Set permissions
-            val file = File(application.filesDir.toString() + "/cjdroute")
-            file.setExecutable(true)
-        } catch (e: IOException) {
-            e.printStackTrace()
-            Log.e(LOGTAG, "Failed to copy cjdroute file", e)
         }
         //Create and initialize conf file
-        util!!.initializeCjdrouteConfFile()
+        if (!File(application.filesDir.toString()+"/"+AnodeUtil().cjdrouteConfFile).exists()) {
+            util!!.initializeCjdrouteConfFile()
+        }
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -68,22 +79,17 @@ class MainActivity : AppCompatActivity() {
         util = AnodeUtil()
         val prefs = getSharedPreferences("com.anode.anode", Context.MODE_PRIVATE)
 
-        if (prefs.getBoolean("firstrun", true)) {
-            InitializeApp()
-            prefs.edit().putBoolean("firstrun", false).commit()
-        }
-        var buttonStartTest: FloatingActionButton? = findViewById(R.id.button_start_test)
+        //Initialize the app by copying cjdroute and generating the conf file
+        initializeApp()
 
-        buttonStartTest?.setOnClickListener(View.OnClickListener { view ->
-            Snackbar.make(view, "Launching cjdroute...", Snackbar.LENGTH_LONG)
-                    .setAction("Action", null).show()
-            val intent = VpnService.prepare(applicationContext)
-            if (intent != null) {
-                startActivityForResult(intent, 0)
-            } else {
-                onActivityResult(0, Activity.RESULT_OK, null)
-            }
-        })
+        /* We may need the first run check in the future... */
+        /*
+        if (prefs.getBoolean("firstrun", true)) {
+            prefs.edit().putBoolean("firstrun", false).commit()
+        } */
+
+        //Launch CJDROUTE
+        //AnodeUtil().launch()
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean { // Inflate the menu; this adds items to the action bar if it is present.
@@ -105,13 +111,6 @@ class MainActivity : AppCompatActivity() {
             val intent = Intent(this, AnodeVpnService::class.java)
             startService(intent)
         }
-    }
-
-    fun showMessageSnackbar(msg: String?) {
-        var buttonStartTest: FloatingActionButton? = null
-        val view = findViewById<View>(buttonStartTest!!.id)
-        Snackbar.make(view, msg!!, Snackbar.LENGTH_LONG)
-                .setAction("Action", null).show()
     }
 
     companion object {
